Added theme and basic imported content

+++ Makefile
@@ -1,3 +1,4 @@
+PY=python
 PELICAN=pelican
 PELICANOPTS=
 
@@ -6,33 +7,60 @@ INPUTDIR=$(BASEDIR)/content
 OUTPUTDIR=$(BASEDIR)/output
 CONFFILE=$(BASEDIR)/pelicanconf.py
 PUBLISHCONF=$(BASEDIR)/publishconf.py
-HTMLDIR=/var/www/html
+
+SSH_HOST=skaia.csh.rit.edu
+SSH_PORT=22
+SSH_USER=root
+SSH_TARGET_DIR=/var/www/html
 
 help:
-	@echo 'Makefile for a pelican Web site                                           '
-	@echo '                                                                          '
-	@echo 'Usage:                                                                    '
-	@echo '   make html                        (re)generate the web site             '
-	@echo '   make clean                       remove the generated files            '
-	@echo '   make regenerate                  regenerate files upon modification    '
-	@echo '   make publish                     move web pages to production location '
-	@echo '                                                                          '
+	@echo 'Makefile for a pelican Web site                                        '
+	@echo '                                                                       '
+	@echo 'Usage:                                                                 '
+	@echo '   make html                        (re)generate the web site          '
+	@echo '   make clean                       remove the generated files         '
+	@echo '   make regenerate                  regenerate files upon modification '
+	@echo '   make publish                     generate using production settings '
+	@echo '   make serve                       serve site at http://localhost:8000'
+	@echo '   make devserver                   start/restart develop_server.sh    '
+	@echo '   make stopserver                  stop local server                  '
+	@echo '   ssh_upload                       upload the web site via SSH        '
+	@echo '   rsync_upload                     upload the web site via rsync+ssh  '
+	@echo '                                                                       '
 
 
-html: clean
+html: clean $(OUTPUTDIR)/index.html
+
+$(OUTPUTDIR)/%.html:
 	$(PELICAN) $(INPUTDIR) -o $(OUTPUTDIR) -s $(CONFFILE) $(PELICANOPTS)
-	-find output -name ".webassets-cache" | xargs rm -rf
-	-rm output/index2.html
-	@echo 'Done'
 
 clean:
-	find $(OUTPUTDIR) -mindepth 1 -delete
+	[ ! -d $(OUTPUTDIR) ] || find $(OUTPUTDIR) -mindepth 1 -delete
 
 regenerate: clean
 	$(PELICAN) -r $(INPUTDIR) -o $(OUTPUTDIR) -s $(CONFFILE) $(PELICANOPTS)
 
-#publish: html
-publish:
-	rsync -aqz $(OUTPUTDIR)/ $(HTMLDIR)
+serve:
+	cd $(OUTPUTDIR) && $(PY) -m pelican.server
 
-.PHONY: help html clean regenerate publish
+devserver:
+	$(BASEDIR)/develop_server.sh restart
+
+stopserver:
+	kill -9 `cat pelican.pid`
+	kill -9 `cat srv.pid`
+	@echo 'Stopped Pelican and SimpleHTTPServer processes running in background.'
+
+publish:
+	$(PELICAN) $(INPUTDIR) -o $(OUTPUTDIR) -s $(PUBLISHCONF) $(PELICANOPTS)
+
+ssh_upload: publish
+	scp -P $(SSH_PORT) -r $(OUTPUTDIR)/* $(SSH_USER)@$(SSH_HOST):$(SSH_TARGET_DIR)
+
+rsync_upload: publish
+	rsync -e "ssh -p $(SSH_PORT)" -P -rvz --delete $(OUTPUTDIR)/ $(SSH_USER)@$(SSH_HOST):$(SSH_TARGET_DIR) --cvs-exclude
+
+dropbox_upload: publish
+	cp -r $(OUTPUTDIR)/* $(DROPBOX_DIR)
+
+.PHONY: html help clean regenerate serve devserver publish ssh_upload rsync_upload
